Rename the user route file
